pyproject.toml
[project]
name = "openalex-search"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "boto3>=1.35.63",
    "fastparquet>=2024.11.0",
    "ipykernel>=6.29.5",
    "openai>=1.54.4",
    "pandas>=2.2.3",
    "pgvector>=0.3.6",
    "psycopg2-binary>=2.9.10",
    "pydantic>=2.9.2",
    "python-dotenv>=1.0.1",
    "requests>=2.32.3",
    "sqlmodel>=0.0.22",
]
